Add functionality to view a single garden by passing props through App
Add createdGardenIndex to api in order to add createdGardens as a piece
of state in App.js. Add viewGarden to App state and setViewGarden
to App to so they are passed as props to the GardenIndex component,
and that state, once set, is used to view a single garden in
GardenView.